refactor(LeaderboardHeader): left, right 삭제
