+++ app/src/Leaderboard/pages/Comment/index.tsx
@@ -4,9 +4,7 @@ import { useSearchParams } from 'react-router-dom';
 
 import { leaderboardArgsAtom } from '@/Leaderboard/atoms/leaderboardArgsAtom';
 import { Leaderboard } from '@/Leaderboard/components/Leaderboard';
-import LeaderboardDateDescriptor from '@/Leaderboard/components/Leaderboard/LeaderboardDateDescriptor';
 import LeaderboardHeader from '@/Leaderboard/components/Leaderboard/LeaderboardHeader';
-import PromoSelectList from '@/Leaderboard/components/PromoSelect/PromoSelectList';
 import { LeaderboardResultSkeleton } from '@/Leaderboard/components/skeletons/LeaderboardResultSkeleton';
 import {
   LEADERBOARD_DEFAULT_OPTIONS,
@@ -18,15 +16,8 @@ import { DateTemplate } from '@shared/__generated__/graphql';
 import { FullPageApolloErrorView } from '@shared/components/ApolloError/FullPageApolloErrorView';
 import { Pagination } from '@shared/components/Pagination';
 import { Seo } from '@shared/components/Seo';
-import {
-  DeferredComponent,
-  Select,
-  SelectContent,
-  SelectTrigger,
-  VStack,
-} from '@shared/ui-kit';
+import { DeferredComponent, VStack } from '@shared/ui-kit';
 import { useDeviceType } from '@shared/utils/react-responsive/useDeviceType';
-import { leaderboardPromoListAtom } from '@/Leaderboard/atoms/leaderboardPromoListAtom';
 
 import { GET_LEADERBOARD_COMMENT } from './queries/getLeaderboardComment';
 
@@ -36,14 +27,9 @@ export default function LeaderboardCommentPage() {
   const { PAGE, PROMO } = LEADERBOARD_PARAM_KEYS;
 
   const leaderboardArgs = useAtomValue(leaderboardArgsAtom);
-  const promoList = useAtomValue(leaderboardPromoListAtom);
 
   if (leaderboardArgs === null) {
     throw new Error('leaderboardArgs is null');
-  }
-
-  if (promoList === null) {
-    throw new Error('promoList is null');
   }
 
   const { loading, error, data } = useQuery(GET_LEADERBOARD_COMMENT, {
@@ -116,25 +102,10 @@ export default function LeaderboardCommentPage() {
       <VStack w="100%" spacing="6rem">
         <VStack w="100%" spacing="2rem">
           <LeaderboardHeader
-            left={
-              <Select
-                width="21rem"
-                onValueChange={handlePromoChange}
-                defaultValue={promo !== null ? String(promo) : undefined}
-                defaultRenderValue={promo !== null ? `${promo}기` : undefined}
-              >
-                <SelectTrigger placeholder="전체" />
-                <SelectContent maxHeight="20rem">
-                  <PromoSelectList list={promoList ?? []} />
-                </SelectContent>
-              </Select>
-            }
-            right={
-              <LeaderboardDateDescriptor
-                start={new Date(start)}
-                end={new Date(end)}
-              />
-            }
+            currPromo={promo}
+            onPromoChange={handlePromoChange}
+            start={new Date(start)}
+            end={new Date(end)}
           />
           <Leaderboard me={me} list={nodes} unit={unit} />
         </VStack>
